requirements.txt
### requirements.txt
scapy
python-nmap
rich
